# ENH: vectorizing float32 implementation of np.exp & np.log