… numpy/core/code_generators/generate_umath.py