@juliantaylor 2019-03-18
This looks very nice and uses our runtime detection so automatic performance improvement for users with the right hardware using gcc.

It is always a bit dodgy to implement the basic functions ourselves and not using libm for it as we lose out on potential improvements there (e.g. glibc has avx512 code for the these functions and more).
Then again we gain more control over when they change and are more reliable over different platforms.

Would be interesting to test this against a few other projects like scipy and see if tests break.