@juliantaylor 2019-03-20
Do you have a reference for the algorithm used here?