@juliantaylor 2019-03-20
I would like to compare this glibc's exp implementation (there is an old branch of mine which allows using it from numpy https://github.com/numpy/numpy/pull/7865) but mostly just out of curiosity.

A concern is that we do not have good coverage if these basic math functions actually compute correct results over the full range, while I trust it is correct currently to keep it correct coverage would be good.
Then again the same thing applies to our fallback implementations in case the system libm does not provide decent functions... so it is not any worse than the status quo if we merge as is.