@r-devulap 2019-03-21
> Do you have a reference for the algorithm used here?

The algorithm is based of the book "Elementary Functions: Algorithms and Implementation" by Jean-Michel Muller. Basically algorithm for transcendental functions is broken into 3 steps: 

1. Range reduction (chapter 9 talks about this for sin, cos, exp and log). I ended up using the Cody-Waite range reduction method which is very SIMD friendly. 
2. Polynomial approximation in the reduced range (using a mini-max polynomial approximation), the coefficients of which can be computed by solving Remez's algorithm. [Boost library](https://www.boost.org/doc/libs/1_67_0/libs/math/doc/html/math_toolkit/remez.html) provides a nice solver for this and I used it to experiment with different polynomials and used one that provided reasonable accuracy.  
3. Reconstruction of the original result, which is a simple multiplication or addition in this case.