@r-devulap 2019-03-21
> I would like to compare this glibc's exp implementation (there is an old branch of mine which allows using it from numpy #7865) but mostly just out of curiosity.
> 
> A concern is that we do not have good coverage if these basic math functions actually compute correct results over the full range, while I trust it is correct currently to keep it correct coverage would be good.
> Then again the same thing applies to our fallback implementations in case the system libm does not provide decent functions... so it is not any worse than the status quo if we merge as is.

I did compare the accuracy of this implementation with glibc's libmvec library (32-bit version) and turns out libmvec has a slightly worse accuracy that this implementation. 

| Function | libmvec max ULP | this patch max ULP | 
| -- | -- | -- |
| sine | 2.47 | 1.49 |
| cosine | 2.39 | 1.49 |
| exp | 2.93 | 2.52 |
| log | 3.96 | 3.83 |

I have another patch ready which implements float32 implementation of sine and cosine using AVX2 and AXV512 with a max ULP of 1.49. If you like, I can add that to this pull request as well.