@juliantaylor 2019-03-21
very nice, can you provide the code you used to compare the implementations?

please do not add sin/cos to this PR, lets do that one at the time.