@r-devulap 2019-03-21
> very nice, can you provide the code you used to compare the implementations?
> 
> please do not add sin/cos to this PR, lets do that one at the time.

The code is fairly simple actually: 

Have a for loop iterating over all 32 bit floats and for each float: 

1. `double my_value = (double) AVX2_exp_FLOAT(); // upscale the float to a double`
2. `double reference_value = exp(); // glibc's scalar function `
3. `double ULP_error  = compute_ULP(my_value, reference_value); // a small helper function which computes the error is terms of ULP`  

you can then compute the max ULP error and do the same to get max relative error.