@mattip 2019-03-27
There are warnings when compiling with various options, are some of the #ifdef clauses not quite correct?