@r-devulap 2019-03-27
> There are warnings when compiling with various options, are some of the #ifdef clauses not quite correct?

Could you please tell me which compiler and flags you are using? With gcc 7.3.0, it seems to build fine. There are some warnings, but nothing seems to be from this patch. 

If you are using clang6.0.0, then this patch fails to build because of this error (gcc or even clang7.0 builds fine): 

_configtest.c:7:20: error: instruction requires: AVX-512 ISA                                                                                                                                  
`  __asm__ volatile("vpaddd %zmm1, %zmm2, %zmm3");`