@mattip 2019-03-27
The CI is failing. You should see a red X next to the changeset label above:
```
BUG: Fixing compile time error for clang                               X f802e92 
```
Clicking on that X should open a set of links to failing CI runs: [travis ci](https://travis-ci.org/numpy/numpy/builds/510032107?utm_source=github_status&utm_medium=notification),  and [azure macos and 32 bit linux](https://dev.azure.com/numpy/numpy/_build/results?buildId=2556&view=logs). The CI fails on compiler warnings.