@r-devulap 2019-03-27
> Linux 32 bit is [failing](https://dev.azure.com/numpy/numpy/_build/results?buildId=2607&view=logs&jobId=1cdee7e3-4100-5086-4c59-ed0352998504) - `ImportError: /numpy/build/testenv/lib/python3.6/site-packages/numpy/core/_multiarray_umath.cpython-36m-i386-linux-gnu.so: undefined symbol: AVX512F_exp_FLOAT`

yeah, I still don't fully understand why that is failing. Do you have any suggestions?