@tylerjereddy 2019-03-27
Is it worth trying `-march=skylake-avx512` or similar compile flag for the 32-bit Ubuntu docker container? Maybe 32-bit linux requires a little more convincing to emit the appropriate code / use appropriate libs on that platform?