@tylerjereddy 2019-03-27
Even if that did work, I guess you'd still need some kind of guard so that you can still import NumPy on a "regular" compile without the -m flag on 32-bit.

Other random thought--that Docker container will likely have pretty minimal libs installed so if you're expecting something specific it could be missing.