@r-devulap 2019-03-27
hmm, I don't have any dependencies on other libs. If GCC supports avx512f instrinsics, then it compiles it. At the time of loading .so file (when importing numpy), if cpuid/xgetbv returns a successful check for avx512f, then it points to the avx512f version of the function FLOAT_exp_avx512f, which is where the failure seems to occur.