@mattip 2019-03-27
fired up my bionic 32 bit chroot. Forcing the archtecture with ``CFLAGS='-march=skylake-avx512'` indeed allows the module to load, but then it crashes with `Illegal instruction` when running tests. Apparently the check is not quite good enough for a chroot.